requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.1
django-allauth==0.58.2
python-dotenv==1.0.0
requests==2.31.0
django-environ==0.11.2
django-widget-tweaks==1.5.0
django-bootstrap5==23.3
django-embed-video==1.4.9